requirements.txt
PyQt6
pyqtgraph
pymodbus
pyyaml
